Fix possible loss of data warning
warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data

Amends merge resolution of 24d0266ee45cf6a3c5b9142453966199702fbf90.

+++ src/qml/memory/qv4mm.cpp
@@ -733,7 +733,7 @@ MemoryManager::MemoryManager(ExecutionEngine *engine)
 
 #ifdef MM_STATS
 static int allocationCount = 0;
-static int lastAllocRequestedSlots = 0;
+static size_t lastAllocRequestedSlots = 0;
 #endif
 
 Heap::Base *MemoryManager::allocString(std::size_t unmanagedSize)
